requirements.txt
# Web framework
fastapi
uvicorn

# Validation / schema
pydantic==2.12.5

# LangChain ecosystem
langchain==1.2.0
langchain-core==1.2.5
langchain-google-genai==4.1.2
langgraph==1.0.5

# Env config
python-dotenv
